requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Werkzeug==2.3.7
qrcode==7.4.2
Pillow==10.0.1
